Update Quagga configuration in barcode.html and barcode.js
In barcode.html, the jQuery script reference has been commented out in line 7 and the source for quagga.min.js has been removed in line 8. Two input fields, with ids chbx_singleChannel and slct_locatorPatchSize have been enabled. In barcode.js, additional parameters for configuring the Quagga barcode reader have been added. These include the size of the input stream in line 14, whether to use single channel mode in line 15, the patch size of the locator in line 28, and whether or not to use half-sample in line 29. Additionally, an onProcessed function has been added in line 33 and the scannedDataCopy function has been modified in lines 80-89 to write the scanned data to the user's clipboard in a newline separated string.